Move the fileprovider mock to its own file and add a doCompletion unit test
